refactor: extract ApplicationCard from page file
